Correct option retrieval
`array_key_exists()` throws an error, if option `padb2_settings` has not been set before (because the variable `$options` then equals `false`, whereas the function expects an array as the second parameter).
On the other hand side, `get_option()` takes a second parameter, that embodies the default value for the option to be retrieved. Therefore it is safe to simply return `$options`.